Move AssignPendingDNSInfo to netfx file
